Optionally ignore images.
Changes pulled from contributor royn5618. There were issues with the
images in his file. He did not need the images, and opted to just ignore
them. This may be a useful option for anyone wanting to extract text
from a docx with embedded image-like-but-not-really embedded files
(e.g., a Visio chart).