feat: preserve newlines in replace_docx_text
replace_docx_text now preserves newlines in the replacement text, so "some string" can be replaced with "some\\nstring". Newlines are written in the docx xml as <br> elements, not paragraphs. This is similar to the soft breaks inserted in MS Word when pressing <Shift-Enter>. Soft breaks will, as before, be exported as "\\n".

closes issue #41 Add breakline when using replace_docx_text method