fix: ignore first row vMerge flags in tbl
Dropdown lists in table cells are written into the xml as nested tables
with a vMerge request in the first row. Docx2Python attempted to copy
values from a previous row to satisfy such vMerge requests. This failed,
because there was no previous row from which to copy. Docx2Python now
ignores vMerge requests in the first row of any table.

closes docx2python/issues/73